feat(codegen): allow to skip files
this has no real usage at current codebase, but in a future PR
I'm going to upgrade to the latest cimgui which adds freetype support.

I'll want to exclude this freetype support from go generation,
as it adds more problems and we don't need it right now.

The easiest way to achive this is just disabling one particular file
but that feature sounds useful in general for the generator